% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{create_gitlab_ci}
\alias{create_gitlab_ci}
\title{Create gitlab continuous integration example}
\usage{
create_gitlab_ci()
}
\arguments{
\item{par.}{}
}
\value{
NULL.
}
\description{
\code{create_gitlab_ci} creates an examples of a continuous integration file using
gitlab.
}
\author{
Erick A. Chacón-Montalván
}
