% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{create_proj}
\alias{create_proj}
\title{Create folders and files for the project}
\usage{
create_proj(
  dir = getwd(),
  yaml_file = system.file("reprodown.yaml", package = "reprodown")
)
}
\arguments{
\item{dir}{Directory where to create our project. By default, this is the current
working directory.}

\item{yaml_file}{A file that defines the structure of the project.}
}
\description{
\code{create_proj} creates folders based on the structure of a "yaml_file". By default
it is the structure of the file \code{system.file("reprodown.yaml", package =
"reprodown")}.
}
\details{
For the yaml_file. Each bullet item is understood as a folder, its corresponding
sub-items are considered sub-folders.
}
\author{
Erick A. Chacón-Montalván
}
