% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprodown.R
\name{make_dependency}
\alias{make_dependency}
\title{Create Dependency Rule Based on Yaml Header}
\usage{
make_dependency(yaml_list)
}
\arguments{
\item{yaml_list}{A yaml list obtained from the yaml header of a Rmd file.}
}
\value{
A string with the Makefile syntax of the dependency.
}
\description{
\code{make_dependency} Creates a string with the Makefile syntax to create a
dependency rule based on a yaml list.
}
\details{
It evaluates the yaml list to obtain the prerequisites (a parameter starting with
"prer") and the targets (a parameter defined with "targ").
}
\author{
Erick A. Chacón-Montalván
}
