% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprodown.R
\name{prepare_yaml}
\alias{prepare_yaml}
\title{Prepare Default Targets and Prerequisites for Makefile}
\usage{
prepare_yaml(file_rmd, dir_blog = "docs")
}
\arguments{
\item{file_rmd}{File to obtain the initial yaml header. This should be relative to
the project root path.}

\item{dir_blog}{File to obtain the initial yaml header.}
}
\value{
A yaml list with added default target and prerequisite.
}
\description{
\code{prepare_yaml} Obtains the yaml header of a Rmd document to add the default
prerequisites and targets to be used in a Makefile.
}
\details{
The yaml header is checked to add the default target "html" file and the default
prerequisite "Rmd" file. It append these default in case other targets and
prerequisites have already been defined in the header file. These can be defined
using a parameter starting with "targ" (e.g. targets: path/to/file.RData) for
targets and a parameter starting with "prer" (e.g. prerequisites:
path/to/req1.RData path/to/req2.RData) for prerequisites.
}
\author{
Erick A. Chacón-Montalván
}
