% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprodown.R
\docType{package}
\name{reprodown}
\alias{reprodown}
\title{reprodown: Reproducible Projects with R Markdown and Hugo}
\description{
The \code{reprodown} package provide tools to improve reproducibility of your
project and to create a custom website of this. It integrates blogdown (a package
to create websites using Hugo and R Markdown) with Makefile (GNU make software to
automatise compilation).
}
\author{
Erick A. Chacón-Montalván
}
