% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefile.R
\name{makefile}
\alias{makefile}
\title{Create Makefile for Rmd Files}
\usage{
makefile(dir_src = "scripts", dir_blog = "docs", command = "")
}
\arguments{
\item{dir_src}{Directory path of the Rmd scripts.}

\item{dir_blog}{Directory path of your \code{blogdown} web inside your project.}

\item{command}{Command to build rmd files to html.}
}
\value{
A Makefile with the rules
}
\description{
\code{make_file} Creates a Makefile based on the prerequisites and targets defined
on the yaml header of Rmd scripts.
}
\details{
The prerequisites and targets of a Rmd files are defined on the yaml header using
parameters starting with "prer" or "targ" respectively. It is not required to put
the obvious prerequisite (the same Rmd file) and the obvious target (the
associated html file). An example of a yaml header is below:\cr
---\cr
title: Causal Analysis of COVID-19\cr
prerequisites:\cr
     - data/counts.RData\cr
     - data/transportation.RData\cr
targets:\cr
     - data/model/spatial.RData\cr
     - data/model/global.RData\cr
---
}
\author{
Erick A. Chacón-Montalván
}
